+++ myMY/__init__.py
@@ -102,8 +102,8 @@ class About():
         return str(self.version)
 
 
-systemInfoObject = About(version=0.493, status='Beta Release',
-                         build=20231227, version_note='Data export draft')
+systemInfoObject = About(version=0.5, status='Beta Release',
+                         build=20231228, version_note='Deletion system enhancement')
 systemInfo = systemInfoObject.__str__()
 systemVersion = systemInfoObject.getSystemVersion()