+++ CHANGELOG.md
@@ -23,3 +23,4 @@ Created on Dec 20, 2023
 - Beta Release: Transaction DB model adjustment + add migration on December 26, 2023 -> **0.491**
 - Beta Release: Changed transaction values on December 26, 2023 -> **0.492**
 - Beta Release: Data export draft on December 27, 2023 -> **0.493**
+- Beta Release: Deletion system enhancement on December 28, 2023 -> **0.5**